refactor: move type definitions into separate types file
